+++ Makefile
@@ -9,7 +9,8 @@ SRC_DIR := src
 all: $(NAME)
 
 $(NAME): developing
-	cc src/*.c -Ilib/libft/inc/ -Llib/libft/lib -lft -lreadline -o minishell -g -Wall -Wextra
+	cc src/*.c src/lexical_analyzer/*.c \
+	-Ilib/libft/inc/ -Llib/libft/lib -lft -lreadline -o minishell -g -Wall -Wextra
 
 developing:
 	@echo "Compiling..."
